feat: Enhance analysis progress handling and result serialization; support decimal progress values and improve data structure for completion events

+++ ai-service/models/response_models.py
@@ -162,7 +162,7 @@ class LegacyAnalysisResponse(BaseModel):
 
 class ProgressUpdate(BaseModel):
     analysis_id: str
-    progress: int
+    progress: float  # Changed from int to float to handle decimal progress values
     status: "AnalysisStatus"
     current_step: str
     timestamp: datetime
